backend-deprecated/requirements.txt
asgiref==3.6.0
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.1.10
django-cors-headers==3.14.0
django-environ==0.11.2
djangorestframework==3.14.0
gunicorn==21.2.0
idna==3.6
packaging==23.2
pipenv==2023.3.20
psycopg2-binary==2.9.9
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.1.0
virtualenv==20.21.0
virtualenv-clone==0.5.7
